Initial Commit
Project is building with libigl, Directional, and libQEx as
submodules.

Created a new QuadMesh class which is similar to the previous
Mesh class, but specialized to just quads. This allows us to
simplify some methods, and enables simple and fast implementation
of dual loops and edge loops.

Another difference is that this implementation makes heavy use
of libigl, whereas the previous mesh class avoided it.