Fixed edge labels for quad mesh.
Fixed edge labels so they now appear on both sides of an edge.
Quad vertex labels have been removed since they had a transparency
bug that I have not been able to fix.

There is an outstanding bug in edge labels in that they are not
visible through the dual-edge labels. This is a minor bug so it
will not be fixed for now. Interestingly, half edges do not exhibit
this behavior.