Added projection of glyphs onto a surface
Mostly works, but is a bit artifacty, and has trouble around the
armpits of the sweater.

Changes that should happen:
LUV for labeled quad mesh should natively be nx3 with a column
of 1s - this would avoid the necessity of constantly padding with
a column of 1s in many places.

Better subdivision! Using the false barycentric coordinates will
give us slivers in the long-run.

Tweak offset distances - these may need to be larger on curved
surfaces.

If we could use the original mesh UVs that were used to generate
the quad mesh in order to aid in back-projection, that might help
a lot around saddle points like armpits.