Moved coarse knit generation to coarse graph
Re-wrote coarse knit graph generation and moved it to live in the
coarse mesh class as the method "get_dual()". The reason is that the
graph formulation and the patches are agnostic to the type of polygon
they are embedded in, so having the graph class depend on the mesh
class created an unneccessary dependency. This way if we want to
start supporting T-Junctions or polymeshes in the future we only need
to replace the mesh class.