Added knit graph vis. + missing geo constraints
Added a knit graph visualizer to the labeling UI.

Found that the quad and edge geo constraints were not being added.
Added them, found a bug in quad constraints (side identification
was wrong). Appears to work now for small / simple models, but
runs forever on larger models (the skirt). However, this model
is smaller than ones we ran in the previous submission, so we
likely need to adopt some of the old optimizations again, or else
there is another bug elsewhere.